+++ lib/js_regex/error.rb
@@ -1,0 +1,5 @@
+class JsRegex
+  # This is mixed into errors, e.g. those thrown by the parser,
+  # allowing to `rescue JsRegex::Error`.
+  module Error; end
+end