+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.7.0] - 2021-03-03
+
+### Added
+
+- added `JsRegex::Error` as a mixin to all errors (mostly parser errors) to simplify catching them
+- added `JsRegex::new!` as a strict method that will raise if the input can not be fully converted
+
 ## [3.6.0] - 2020-12-02
 
 ### Added